Removed the warnings from chat services.cs

+++ src/DBServices/ChatServices.cs
@@ -19,44 +19,52 @@ public class ChatServices {
         var conversation = new Conversation(Guid.NewGuid().ToString(), name);
         foreach (var username in usernames)
         {
-            var user = _context.Users.SingleOrDefault(u => u.Username == username);
+            var user = _context.Users!.SingleOrDefault(u => u.Username == username);
             if(user!=null) {
             conversation.Users.Add(user);
             }
         }
-        _context.Conversations.Add(conversation);
+        _context.Conversations!.Add(conversation);
         _context.SaveChanges();
         return conversation;
     }
     public void AddUserToConversation(string conversationID, string userID)
     {
-        var conversation = _context.Conversations.Find(conversationID);
-        var user = _context.Users.Find(userID);
-        conversation.Users.Add(user);
-        _context.SaveChanges();
+        var conversation = _context.Conversations!.Find(conversationID);
+        var user = _context.Users!.Find(userID);
+            if(conversation != null && user != null){
+                conversation!.Users.Add(user);
+                _context.SaveChanges();
+            }
     }
     public void RemoveUserFromConversation(string conversationID, string userID)
     {
-        var conversation = _context.Conversations.Find(conversationID);
-        var user = _context.Users.Find(userID);
-        conversation.Users.Remove(user);
-        _context.SaveChanges();
+        var conversation = _context.Conversations!.Find(conversationID);
+        var user = _context.Users!.Find(userID);
+        if(conversation != null && user != null){
+            conversation.Users.Remove(user);
+            _context.SaveChanges();
+        }
     }
     public void DeleteConversation(string conversationID)
     {
-        var conversation = _context.Conversations.Find(conversationID);
-        _context.Conversations.Remove(conversation);
-        _context.SaveChanges();
+        var conversation = _context.Conversations!.Find(conversationID);
+        if(conversation != null){
+            _context.Conversations.Remove(conversation);
+            _context.SaveChanges();
+        }
     }
     public void SendMessage(string body, string conversationID, string senderID)
     {
         var message = new Message(Guid.NewGuid().ToString(),body, conversationID, senderID, DateTime.Now, false);
-        _context.Messages.Add(message);
-        _context.SaveChanges();
+        if(message != null){
+            _context.Messages!.Add(message);
+            _context.SaveChanges();
+        }
     }
     public List<Message> GetMessages(string conversationID)
     {
-        List<Message> messages = _context.Messages.Where(m => m.ConversationID == conversationID).ToList();
+        List<Message> messages = _context.Messages!.Where(m => m.ConversationID == conversationID).ToList();
         foreach (var message in messages)
         {
             message.Sent = true;
@@ -66,7 +74,7 @@ public class ChatServices {
     }
     public List<Conversation> GetConversations(string userID)
     {
-        List<Conversation> conversations = _context.Conversations.Where(c => c.Users.Any(u => u.UserId == userID)).ToList();
+        List<Conversation> conversations = _context.Conversations!.Where(c => c.Users.Any(u => u.UserId == userID)).ToList();
         return conversations;
     }
     public void Dispose()
